# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Learning Platform API",
        "description": "API documentation for the Online Learning Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get all sessions for the authenticated user",
                "operationId": "0b722a0b8a07394301d29492524e35c3",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sessions for the authenticated user"
                    }
                }
            }
        },
...
        "/api/sessions/{sessionId}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get session details",
                "operationId": "c52f6cbf354c88760b813c01770a7aef",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session details"
                    }
                }
            }
        },
...
    },
...
}